сделал механику износа скинов, починил добавление xp при открытии кейса

+++ app/src/main/java/com/ezDrop/app/data/db/dao/UserDao.kt
@@ -63,5 +63,5 @@ interface UserDao {
     suspend fun updateAvatarUri(userId: Long, uri: String?): Int
 
     @Query("UPDATE users SET netWorth = :netWorth WHERE id = :userId")
-    suspend fun updateNetWorth(userId: Long, netWorth: Int): Int
+    suspend fun updateNetWorth(userId: Long, netWorth: Long): Int
 }
